add lesson on file permissions
Simplified sentence on 'd' file type.

What do the letters really mean?

Formatted to make "r", "w", "x" and "-" stand out. Added brief supporting examples.

Section on users and groups.

Brief section on users and groups with an example. Raised awareness of 'sudo' so its mention of it is more meaningful during supercomputer training.

List to support permissions.png.

Summarised information contained in 'permissions.png' using dot points (carry over from unused section).

Formatting in "Changing Permissions", syntax and example.

Introduced the syntax of chmod as it is a little different from the other commands that they have encountered so far. Tweaked the example to be indicative of chmod usage in the context of a Pawsey project.

deleted chmod examples

Removed the 'complicated' examples as I'm not sure that they fit within the context of an introductory workshop.

Group on pawsey systems.

Mentioned that interns will be in a group with their supervisors and other collaborators.

Deleted lines starting with ">".

Linux

Removed explicit mention of Linux, need to be clear that this information is more-or-less general to Unix-like OS.

spelling and grammer

Swapped list and images.

The output contains directories whereas the image does not. I think the image is more at the end of this block in general (though I'm not certain about this).

formatting to match other lists.

Revert "Deleted lines starting with ">"." as they are actually required
for exercises

This reverts commit d144d6c65ddab28dd97d1c2aecaa0c70577756ee.